reorganize repo structure (cont)
