+++ keycastow.vcxproj
@@ -24,17 +24,17 @@
     <ClCompile Include="src\keycast\MyStaticVar.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="resource.h" />
+    <ClInclude Include="res\resource.h" />
     <ClInclude Include="src\keycast\keycast.h" />
     <ClInclude Include="src\keycast\keylog.h" />
     <ClInclude Include="src\keycast\MyStaticVar.h" />
     <ClInclude Include="src\keycast\timer.h" />
   </ItemGroup>
   <ItemGroup>
-    <ResourceCompile Include="keycastow.rc" />
+    <ResourceCompile Include="res\keycastow.rc" />
   </ItemGroup>
   <ItemGroup>
-    <Image Include="keycastow.ico" />
+    <Image Include="res\keycastow.ico" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{D9A196AF-F63D-4E8E-8724-9364FD8EB4D4}</ProjectGuid>
